refactor(transcripts): split migration insert transaction
